rfid_project/requirements.txt
flask
flask-cors
openai-whisper
torch
numpy
# Note: Whisper uses ffmpeg externally; install ffmpeg separately on your system.
